Move Avatar_Privacy\Core\Hasher to Avatar_Privacy\Tools\Hasher
