+++ includes/avatar-privacy/core/class-user-fields.php
@@ -27,7 +27,7 @@
 namespace Avatar_Privacy\Core;
 
 use Avatar_Privacy\Core\API;
-use Avatar_Privacy\Core\Hasher;
+use Avatar_Privacy\Tools\Hasher;
 
 
 /**
