+++ includes/avatar-privacy/data-storage/database/class-comment-author-table.php
@@ -27,8 +27,8 @@
 namespace Avatar_Privacy\Data_Storage\Database;
 
 use Avatar_Privacy\Data_Storage\Database\Table;
-use Avatar_Privacy\Core\Hasher;
 use Avatar_Privacy\Data_Storage\Network_Options;
+use Avatar_Privacy\Tools\Hasher;
 
 /**
  * The database table used for storing (anonymous) comment author data.
