+++ includes/avatar-privacy/upload-handlers/class-custom-default-icon-upload-handler.php
@@ -27,11 +27,12 @@
 namespace Avatar_Privacy\Upload_Handlers;
 
 use Avatar_Privacy\Core;
-use Avatar_Privacy\Core\Hasher;
 use Avatar_Privacy\Core\Settings;
 
 use Avatar_Privacy\Data_Storage\Filesystem_Cache;
 use Avatar_Privacy\Data_Storage\Options;
+
+use Avatar_Privacy\Tools\Hasher;
 
 use Avatar_Privacy\Upload_Handlers\Upload_Handler;
