+++ tests/avatar-privacy/core/class-user-fields-test.php
@@ -33,7 +33,7 @@ use Brain\Monkey\Functions;
 use Mockery as m;
 
 use Avatar_Privacy\Core\User_Fields;
-use Avatar_Privacy\Core\Hasher;
+use Avatar_Privacy\Tools\Hasher;
 
 /**
  * Avatar_Privacy_Factory unit test.
